Refactor VideoRecorderSessionHandler into partial class files
Split the 656-line VideoRecorderSessionHandler.cs into focused partial
files matching the FfmpegVideoRecorder pattern:
- VideoRecorderSessionHandler.cs (wiring + lifecycle)
- VideoRecorderSessionHandler.DataConsumer.cs
- VideoRecorderSessionHandler.VideoProduction.cs
- VideoRecorderSessionHandler.SegmentPruning.cs
- VideoRecorderSessionHandler.Helpers.cs

No logic changes; each file is under 300 lines.

Fixes #9592
